finished events and updated readme

+++ backend/routes/api/events.js
@@ -59,7 +59,14 @@ router.post('/create', (req, res) => {
 router.get('/show', (req, res) => {
     let event_id = req.body.event_id;
     //console.log(event_id);
-    Event.find( {_id: event_id}  ).then((result) => {
+    Event.findOne( {_id: event_id}  ).then((result) => {
+
+      if (!result) {
+        return res.status(400).json({
+          success: false,
+          msg: "Event not found."
+        });
+      }
       res.status(200).send(result);
     })   
 });
@@ -77,52 +84,49 @@ router.post('/add', (req, res) => {
   jwt.verify(token, key, function(err, decoded) {
     if (err) return res.status(500).send({ success: false, message: 'Failed to authenticate token.' });
   
-    let user_id;
-    User.find( {username: req.body.username} ,function(err, doc){
-      console.log(doc);
+    User.findOne( {username: req.body.username} ,function(err, doc){
       
-      this.user_id = doc._id;
-    });   
-
-    if (!user_id) {
-      return res.status(400).json({
-        success: false,
-        msg: "User not found."
+      const user_id = doc._id;
+      if (!user_id) {
+        return res.status(400).json({
+          success: false,
+          msg: "User not found."
+        });
+      }
+      const username = req.body.username;
+      const event_id = req.body.event_id;
+  
+      User.findOneAndUpdate({
+        username: username,
+        events: {$ne: event_id}
+      },
+      { $push: {events: [event_id] }} ,function(err, doc){
+        if(err){
+            console.log("Something wrong when updating data!");
+        }
+        doc ? console.log("Added this event to user's profile!") : console.log("Event already in this user!");
       });
-    }
-
-    const username = req.body.username;
-    console.log(user_id);
-    let event_id = req.body.event_id;
-    console.log(event_id);
-
-    User.findOneAndUpdate({
-      username: username,
-      events: {$ne: event_id}
-    },
-    { $push: {events: [event_id] }} ,function(err, doc){
-      if(err){
-          console.log("Something wrong when updating data!");
-      }
-      doc ? console.log("Added this event to user's profile!") : console.log("Event already in this user!");
+  
+      Event.findOneAndUpdate({
+        _id: event_id,
+        users: {$ne: user_id}
+      },
+      { $push: {users: user_id }},function(err, doc){
+        if(err){
+            console.log("Something wrong when updating data!");
+        }
+        doc ? console.log("Added this user to event!") : console.log("User already has this event!");
+      });
+  
+      return res.status(200).json({
+        success: true,
+        msg: "Congrats, user " + username + " is in event " + event_id
     });
-
-    Event.findOneAndUpdate({
-      _id: event_id,
-      users: {$ne: user_id}
-    },
-    { $push: {users: [user_id] }},function(err, doc){
-      if(err){
-          console.log("Something wrong when updating data!");
-      }
-      doc ? console.log("Added this user to event!") : console.log("User already has this event!");
     });
+    });   
+    
 
-    return res.status(200).json({
-      success: true,
-      msg: "Congrats, user " + username + " is in event " + event_id
-  });
-  });
+    
 });
 
 /**
@@ -138,52 +142,50 @@ router.post('/remove', (req, res) => {
   jwt.verify(token, key, function(err, decoded) {
     if (err) return res.status(500).send({ success: false, message: 'Failed to authenticate token.' });
   
-    let user_id;
-    User.find( {username: req.body.username} ,function(err, doc){
-      console.log(doc);
+    User.findOne( {username: req.body.username} ,function(err, doc){
       
-      this.user_id = doc._id;
+      const user_id = doc._id;
+
+      if (!user_id) {
+        return res.status(200).json({
+          success: false,
+          msg: "User not found."
+        });
+      }
+  
+      const username = req.body.username;
+      const event_id = req.body.event_id;
+  
+      User.findOneAndUpdate({
+        username: username,
+        events: {$eq: event_id}
+      },
+      { $pull: {events: event_id }},function(err, doc){
+        if(err){
+            console.log("Something wrong when updating data!");
+        }
+        doc ? console.log("Removed this event from user's profile!") : console.log("Event not in this user!");
+      });
+  
+      Event.findOneAndUpdate({
+        _id: event_id,
+        users: {$eq: user_id}
+      },
+      { $pull: {users: user_id }} ,function(err, doc){
+        if(err){
+            console.log("Something wrong when updating data!");
+        }
+        doc ? console.log("Removed this user from event!") : console.log("User not in this event!");
+      });
+  
+      return res.status(200).json({
+        success: true,
+        msg: "Congrats, user " + username + " is removed from event " + event_id
+    });
+    });
     });   
 
-    if (!user_id) {
-      return res.status(200).json({
-        success: false,
-        msg: "User not found."
-      });
-    }
-
-    const username = req.body.username;
-    console.log(user_id);
-    let event_id = req.body.event_id;
-    console.log(event_id);
-
-    User.findOneAndUpdate({
-      username: username,
-      events: {$ne: event_id}
-    },
-    { $pop: {events: [event_id] }} ,function(err, doc){
-      if(err){
-          console.log("Something wrong when updating data!");
-      }
-      doc ? console.log("Removed this event from user's profile!") : console.log("Event not in this user!");
-    });
-
-    Event.findOneAndUpdate({
-      _id: event_id,
-      users: {$ne: user_id}
-    },
-    { $pop: {users: [user_id] }},function(err, doc){
-      if(err){
-          console.log("Something wrong when updating data!");
-      }
-      doc ? console.log("Removed this user from event!") : console.log("User not in this event!");
-    });
-
-    return res.status(200).json({
-      success: true,
-      msg: "Congrats, user " + username + " is removed from event " + event_id
-  });
-  });
+    
 });
 
 /**
@@ -198,29 +200,33 @@ router.post('/update', (req, res) => {
   jwt.verify(token, key, function(err, decoded) {
     if (err) return res.status(500).send({ success: false, message: 'Failed to authenticate token.' });
 
-    let {
-      event_id,
-      title,
-      details,
-      start,
-      end
-    } = req.body;
-    //console.log(event_id);
+    let params = {};
 
+    for(let prop in req.body) if(req.body[prop]) params[prop] = req.body[prop];
     Event.findOneAndUpdate({
-      _id: event_id,
+      _id: params.event_id,
     },
-    { title: title, details: details, start: start, end: end},function(err, doc){
+    params ,function(err, doc){
+      if(!doc){
+        return res.status(200).json({
+          success: false,
+          msg: "Event not found!"
+      });
+      }
+      
       if(err){
           console.log("Something wrong when updating data!");
       }
       doc ? console.log("Event updated!") : console.log("Nothing to change!");
+    
+      return res.status(200).json({
+        success: true,
+        msg: "Congrats, event " + params.event_id + " is updated "
+      });
+    
     });
 
-    return res.status(200).json({
-      success: true,
-      msg: "Congrats, event " + doc.title + " is updated "
-  });
+    
   });   
 });
 
@@ -233,25 +239,34 @@ router.post('/delete', (req, res) => {
   let event_id = req.body.event_id;
   //console.log(event_id);
   Event.findOne( {_id: event_id}  ).then((result) => {
+    if(!result){
+      return res.status(200).json({
+        success: false,
+        msg: "Event not found!"
+    });
+    }
     let users = result.users;
     users.forEach(user =>
       User.findOneAndUpdate({
         _id: user._id,
-        events: {$ne: event_id}
+        events: {$eq: event_id}
       },
-      { $pop: {events: [event_id] }} ,function(err, doc){
+      { $pull: {events: event_id }} ,function(err, doc){
         if(err){
             console.log("Something wrong when updating data!");
         }
         doc ? console.log("Removed this event from user's profile!") : console.log("Event not in this user!");
       }));
-});
-  Event.findOneAndDelete( {_id: event_id}  ).then((result) => {
-    res.status(200).json({
-      success: true,
-      msg: "Congrats, event " + result.title + " is deleted and removed from each user"
+
+      Event.findOneAndDelete( {_id: event_id}  ).then((result) => {
+        res.status(200).json({
+          success: true,
+          msg: "Congrats, event " + result.title + " is deleted and removed from each user"
+      });
+      }) 
+
   });
-  })   
+    
 });
